Move all behaviours to separate package
