+++ src/main/java/org/mastodon/grapher/opengl/mamut/MamutViewGrapherOpenGL.java
@@ -24,10 +24,10 @@ import org.mastodon.app.ui.ViewMenu;
 import org.mastodon.app.ui.ViewMenuBuilder.JMenuHandle;
 import org.mastodon.grapher.opengl.PointCloudFrame;
 import org.mastodon.grapher.opengl.PointCloudPanel;
-import org.mastodon.grapher.opengl.overlays.BoxSelectionBehaviour;
-import org.mastodon.grapher.opengl.overlays.ClickSelectionBehaviour;
+import org.mastodon.grapher.opengl.behaviours.BoxSelectionBehaviour;
+import org.mastodon.grapher.opengl.behaviours.ClickSelectionBehaviour;
 import org.mastodon.grapher.opengl.overlays.DataDisplayZoomGL;
-import org.mastodon.grapher.opengl.overlays.FreeformSelectionBehaviourOpenGL;
+import org.mastodon.grapher.opengl.behaviours.FreeformSelectionBehaviourOpenGL;
 import org.mastodon.mamut.MainWindow;
 import org.mastodon.mamut.MamutMenuBuilder;
 import org.mastodon.mamut.ProjectModel;
